Remove handle port conflict
