+++ src/tracer/src/cli/handlers/mod.rs
@@ -1,10 +1,12 @@
 mod info;
 mod init;
+mod terminate;
 mod uninstall;
 mod update;
 
 pub(super) use info::info;
 pub use init::arguments;
 pub(super) use init::init;
+pub(super) use terminate::terminate;
 pub(super) use uninstall::uninstall;
 pub(super) use update::update;
