+++ src/tracer/src/daemon/server/mod.rs
@@ -1,5 +1,4 @@
 mod daemon_server;
 pub use daemon_server::DaemonServer;
 
-pub mod helper;
 pub mod process_monitor;
